requirements.txt
psycopg2
flask
flake8
flake8-quotes
matplotlib
numpy
requests
gunicorn
python-dotenv
git+https://github.com/tomergroisman/postgresql-python.git